requirements.txt
# FastAPI & Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gunicorn>=21.0.0
prometheus-client>=0.17.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Database & ORM
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
alembic>=1.12.0  # For database migrations
pgvector>=0.2.3  # For vector similarity search

# Environment Variables & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI & Embeddings
openai>=1.12.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Telegram Bot
python-telegram-bot>=20.0

# CV Parsing & File Handling
pdfplumber>=0.10.0  # Extract text from PDFs
python-docx>=1.0.0  # Extract text from Word Documents
pillow>=10.0.0  # For image processing

# Web UI & Admin Dashboard
streamlit>=1.28.0
plotly>=5.18.0  # For interactive charts

# Logging & Monitoring
python-json-logger>=2.0.7
structlog>=23.1.0
sentry-sdk>=1.39.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
loguru>=0.7.2  # Enhanced logging capabilities
pythonjsonlogger>=0.1.0  # JSON logging formatter
rich>=13.7.0  # Rich text and beautiful formatting for terminal output

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0
faker>=19.13.0
black>=23.9.0  # Auto-formats Python code
flake8>=6.1.0  # Linter for enforcing best practices
mypy>=1.7.0  # Static type checking
pre-commit>=3.5.0  # Git hooks for code quality

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
